requirements.txt
openai>=1.0.0
websockets>=12.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
SpeechRecognition>=3.10.0
# Skills: docx and xlsx run.py
python-docx>=1.0.0
openpyxl>=3.0.0
# Tests
pytest>=7.0.0
pytest-asyncio>=0.21.0
